LPS-109249 Split LanguageListItem in Editable or not
