+++ modules/apps/depot/depot-web/src/main/resources/META-INF/resources/js/LanguageListItem.es.js
@@ -12,184 +12,21 @@
  * details.
  */
 
-import ClayButton from '@clayui/button';
-import ClayDropDown from '@clayui/drop-down';
-import ClayIcon from '@clayui/icon';
 import ClayLabel from '@clayui/label';
 import ClayTable from '@clayui/table';
-import classNames from 'classnames';
-import React, {useEffect, useRef, useState} from 'react';
-import {useDrag, useDrop} from 'react-dnd';
+import React from 'react';
 
-const noop = () => {};
-
-const DROP_ZONES = {
-	BOTTOM: 'BOTTOM',
-	TOP: 'TOP',
-};
-
-const ACCEPTING_TYPES = {
-	ITEM: 'ITEM',
-};
-
-// Defines the distance in px from the border to the center of an item where
-// dropZone changes from TOP/BOTTOM to ELEMENT
-const ITEM_HOVER_BORDER_LIMIT = 28;
-
-const isValidTarget = (source, target, dropZone) => {
-	return !!(
-		source.id !== target.id &&
-		((dropZone === DROP_ZONES.TOP &&
-			(target.index < source.index || target.index > source.index + 1)) ||
-			(dropZone === DROP_ZONES.BOTTOM &&
-				(target.index > source.index ||
-					target.index < source.index - 1)))
-	);
-};
-
-const getDropZone = (ref, monitor) => {
-	if (!ref.current) {
-		return;
-	}
-
-	const clientOffset = monitor.getClientOffset();
-	const dropItemBoundingRect = ref.current.getBoundingClientRect();
-	const hoverTopLimit = ITEM_HOVER_BORDER_LIMIT;
-	const hoverClientY = clientOffset.y - dropItemBoundingRect.top;
-
-	let dropZone = DROP_ZONES.BOTTOM;
-
-	if (hoverClientY < hoverTopLimit) {
-		dropZone = DROP_ZONES.TOP;
-	}
-
-	return dropZone;
-};
-
-const LanguageListItem = ({
-	displayName,
-	isDefault,
-	localeId,
-	onMakeDefault,
-	showActions,
-	index,
-	onItemDrop = noop,
-}) => {
-	const [active, setActive] = useState(false);
-	const [dropZone, setDropZone] = useState();
-	const ref = useRef();
-
-	const [{isDragging}, drag] = useDrag({
-		collect: monitor => ({
-			isDragging: !!monitor.isDragging(),
-		}),
-		item: {
-			id: localeId,
-			index,
-			type: ACCEPTING_TYPES.ITEM,
-		},
-	});
-
-	const [{isOver}, drop] = useDrop({
-		accept: ACCEPTING_TYPES.ITEM,
-		canDrop(source, monitor) {
-			const dropZone = getDropZone(ref, monitor);
-
-			return isValidTarget(source, {id: localeId, index}, dropZone);
-		},
-		collect: monitor => ({
-			isOver: !!monitor.isOver(),
-		}),
-		drop(source, monitor) {
-			if (monitor.canDrop()) {
-				let newIndex = index;
-
-				if (dropZone === DROP_ZONES.BOTTOM) {
-					newIndex = index + 1;
-				}
-
-				onItemDrop(source.id, newIndex);
-			}
-		},
-		hover(source, monitor) {
-			let dropZone;
-
-			if (isOver && monitor.canDrop()) {
-				dropZone = getDropZone(ref, monitor);
-			}
-
-			setDropZone(dropZone);
-		},
-	});
-
-	useEffect(() => {
-		drag(drop(ref));
-	}, [drag, drop]);
-
-	const makeDefault = event => {
-		setActive(false);
-		onMakeDefault({localeId});
-		Liferay.fire('inputLocalized:defaultLocaleChanged', {
-			item: event.currentTarget,
-		});
-	};
-
-	return (
-		<ClayTable.Row
-			className={classNames('language-list-item', {
-				dragging: isDragging,
-				'drop-bottom': isOver && dropZone === DROP_ZONES.BOTTOM,
-				'drop-top': isOver && dropZone === DROP_ZONES.TOP,
-			})}
-			ref={ref}
-		>
-			<ClayTable.Cell expanded>
-				<div className="autofit-row autofit-row-center">
-					<div className="autofit-col autofit-padded-no-gutters">
-						<div className="language-drag-handler">
-							<ClayIcon symbol="drag" />
-						</div>
-					</div>
-					<div className="align-items-center autofit-col autofit-col-expand autofit-padded-no-gutters flex-row justify-content-start">
-						{displayName}
-						{isDefault && (
-							<ClayLabel className="ml-3" displayType="info">
-								{Liferay.Language.get('default')}
-							</ClayLabel>
-						)}
-					</div>
-				</div>
-			</ClayTable.Cell>
-			{showActions && (
-				<ClayTable.Cell align="right">
-					<ClayDropDown
-						active={active}
-						onActiveChange={setActive}
-						trigger={
-							<ClayButton
-								borderless
-								displayType="secondary"
-								monospaced
-								small
-							>
-								<ClayIcon symbol="ellipsis-v" />
-							</ClayButton>
-						}
-					>
-						<ClayDropDown.ItemList>
-							<ClayDropDown.Item
-								data-value={localeId}
-								key={localeId}
-								onClick={makeDefault}
-							>
-								{Liferay.Language.get('make-default')}
-							</ClayDropDown.Item>
-						</ClayDropDown.ItemList>
-					</ClayDropDown>
-				</ClayTable.Cell>
+const LanguageListItem = ({displayName, isDefault}) => (
+	<ClayTable.Row>
+		<ClayTable.Cell expanded>
+			{displayName}
+			{isDefault && (
+				<ClayLabel className="ml-3" displayType="info">
+					{Liferay.Language.get('default')}
+				</ClayLabel>
 			)}
-		</ClayTable.Row>
-	);
-};
+		</ClayTable.Cell>
+	</ClayTable.Row>
+);
 
 export default LanguageListItem;
