+++ modules/apps/depot/depot-web/src/main/resources/META-INF/resources/js/Languages.es.js
@@ -141,11 +141,11 @@ const Languages = ({
 
 					<LanguagesList
 						defaultLocaleId={customDefaultLocaleId}
+						isEditable
 						locales={customLocales}
 						onItemDrop={handleOnItemDrop}
 						onMakeDefault={handleOnMakeDefault}
 						onOpenManageModal={handleOnModalOpen}
-						showActions
 					/>
 				</>
 			)}
